Ajustes de Modelo de Base de Datos y cambios en html sección Liga.
Se ajusta el modelo de bases de datos con más opciones y se incorporan nuevas consultas para dotar de mayor funcionalidad a la sección Ligas.
Se corrigen los html y se añaden nuevos que comprendan las nuevas acciones posible a realizar (búsquedas por nombre y añadir ligas).

+++ basketGO/src/main/java/es/sidelab/urjc/Equipo.java
@@ -12,7 +12,7 @@ public class Equipo {
 	@GeneratedValue (strategy = GenerationType.AUTO)
 	private long id;
 	@GeneratedValue (strategy = GenerationType.AUTO)
-	private int fichaEquipo;
+	private long fichaEquipo;
 	private String nombreEquipo;
 	private int puntuacion;
 	private int numeroDerrotas;
